DT-55: Fix unlink rights on res partner (Client archivage group)

+++ odoo/local-src/specific_base/models/res_partner.py
@@ -261,24 +261,24 @@ class ResPartner(models.Model):
         string='Flash test done',
         required=True,
         default='n',
-        )
+    )
     socio_professional_category = fields.Selection(
         selection='_get_socio_pro_categ_selection',
         string='Socio-Professional Category'
-        )
+    )
     family_situation = fields.Selection(
         selection='_get_family_situation_selection'
-        )
+    )
     children_nb = fields.Integer(string='# Children')  # or Selection ???
     why_no_buy = fields.Selection(
         selection='_get_why_no_buy_selection'
-        )
+    )
     maximum_budget = fields.Selection(
         selection='_get_maximum_budget_selection'
-        )
+    )
     beauty_institute_attented = fields.Selection(
         selection=[('y', 'Yes'), ('n', 'No')],
-        )
+    )
     comment1 = fields.Text(string='Anecdote/Vacations')
     comment2 = fields.Text(string='Cosmetic habits / Brand used')
     comment3 = fields.Text(string='Other')
@@ -322,7 +322,7 @@ class ResPartner(models.Model):
                 'token': token,
                 'partner_id': self.id,
                 'email': self.email}
-                )
+            )
             # get link to the survey
             url = '%s/%s' % (diagnostic_survey.public_url, token)
 
@@ -352,4 +352,15 @@ class ResPartner(models.Model):
             raise UserError(_(
                 "Impossible to delete a partner linked to a company"
             ))
+        deny_to_delete = (
+            not self.env.user.has_group(
+                'specific_security.group_client_archive'
+            ) and
+            self.env.user.id != 1
+        )
+        if deny_to_delete:
+            raise UserError(_(
+                "You don't have rights to delete partners"
+            ))
+
         return super(ResPartner, self).unlink()
